+++ xmake/modules/private/service/remote_build/server_session.lua
@@ -15,7 +15,7 @@
 -- Copyright (C) 2015-present, TBOOX Open Source Group.
 --
 -- @author      ruki
--- @file        session.lua
+-- @file        server_session.lua
 --
 
 -- imports
@@ -26,30 +26,35 @@ import("core.base.global")
 import("core.base.option")
 import("core.base.hashset")
 import("core.base.scheduler")
-import("utils.archive.extract", {alias = "extract_archive"})
-import("private.service.config")
+import("private.service.server_config", {alias = "config"})
 import("private.service.message")
 import("private.service.remote_build.filesync", {alias = "new_filesync"})
 
 -- define module
-local session = session or object()
+local server_session = server_session or object()
 
--- init session
-function session:init(session_id)
+-- init server session
+function server_session:init(server, session_id)
     self._ID = session_id
+    self._SERVER = server
     local filesync = new_filesync(self:sourcedir(), path.join(self:workdir(), "manifest.txt"))
     filesync:ignorefiles_add(".git/**")
     filesync:ignorefiles_add(".xmake/**")
     self._FILESYNC = filesync
 end
 
--- get session id
-function session:id()
+-- get server session id
+function server_session:id()
     return self._ID
 end
 
--- open session
-function session:open()
+-- get server
+function server_session:server()
+    return self._SERVER
+end
+
+-- open server session
+function server_session:open()
     if self:is_connected() then
         return
     end
@@ -64,8 +69,8 @@ function session:open()
     self:status_save()
 end
 
--- close session
-function session:close()
+-- close server session
+function server_session:close()
     if not self:is_connected() then
         return
     end
@@ -78,17 +83,17 @@ function session:close()
 end
 
 -- set stream
-function session:stream_set(stream)
+function server_session:stream_set(stream)
     self._STREAM = stream
 end
 
 -- get stream
-function session:stream()
+function server_session:stream()
     return self._STREAM
 end
 
 -- diff files
-function session:diff(respmsg)
+function server_session:diff(respmsg)
     local body = respmsg:body()
     vprint("%s: diff files in %s ..", self, self:sourcedir())
 
@@ -137,20 +142,15 @@ function session:diff(respmsg)
 end
 
 -- sync files
-function session:sync(respmsg)
+function server_session:sync(respmsg)
     local body = respmsg:body()
     local stream = self:stream()
     local manifest = assert(body.manifest, "manifest not found!")
     local filesync = self:_filesync()
     local sourcedir = self:sourcedir()
-    local archivefile = os.tmpfile() .. ".zip"
-    local archivedir = archivefile .. ".dir"
+    local archivedir = os.tmpfile() .. ".dir"
     vprint("%s: sync files in %s ..", self, self:sourcedir())
-    if stream:recv_file(archivefile) then
-        vprint("receive archive file, size: %d", os.filesize(archivefile))
-
-        -- extract archive file
-        extract_archive(archivefile, archivedir)
+    if self:_recv_syncfiles(manifest, archivedir) then
 
         -- do sync
         for _, fileitem in ipairs(manifest.inserted) do
@@ -177,20 +177,19 @@ function session:sync(respmsg)
     else
         raise("receive files failed!")
     end
-    os.tryrm(archivefile)
     os.tryrm(archivedir)
     vprint("%s: sync files ok", self)
 end
 
 -- clean files
-function session:clean()
+function server_session:clean()
     vprint("%s: clean files in %s ..", self, self:workdir())
     os.tryrm(self:workdir())
     vprint("%s: clean files ok", self)
 end
 
 -- run command
-function session:runcmd(respmsg)
+function server_session:runcmd(respmsg)
     local body = respmsg:body()
     local program = body.program
     local argv = body.argv
@@ -218,21 +217,17 @@ function session:runcmd(respmsg)
 end
 
 -- get work directory
-function session:workdir()
-    local workdir = config.get("remote_build.server.workdir")
-    if not workdir then
-        workdir = path.join(global.directory(), "service", "remote_build")
-    end
-    return path.join(workdir, "sessons", self:id())
+function server_session:workdir()
+    return path.join(self:server():workdir(), "sessons", self:id())
 end
 
 -- is connected?
-function session:is_connected()
+function server_session:is_connected()
     return self:status().connected
 end
 
 -- get the status
-function session:status()
+function server_session:status()
     local status = self._STATUS
     local statusfile = self:statusfile()
     if not status then
@@ -246,27 +241,27 @@ function session:status()
 end
 
 -- save status
-function session:status_save()
+function server_session:status_save()
     io.save(self:statusfile(), self:status())
 end
 
 -- get status file
-function session:statusfile()
+function server_session:statusfile()
     return path.join(self:workdir(), "status.txt")
 end
 
 -- get sourcedir directory
-function session:sourcedir()
+function server_session:sourcedir()
     return path.join(self:workdir(), "source")
 end
 
 -- get filesync
-function session:_filesync()
+function server_session:_filesync()
     return self._FILESYNC
 end
 
 -- ensure source directory
-function session:_ensure_sourcedir()
+function server_session:_ensure_sourcedir()
     local sourcedir = self:sourcedir()
     if not os.isdir(sourcedir) then
         os.mkdir(sourcedir)
@@ -274,7 +269,7 @@ function session:_ensure_sourcedir()
 end
 
 -- write data from pipe
-function session:_write_pipe(opt)
+function server_session:_write_pipe(opt)
     local buff = bytes(256)
     local wpipe = opt.wpipe
     vprint("%s: %s: writing data ..", self, wpipe)
@@ -295,7 +290,7 @@ function session:_write_pipe(opt)
 end
 
 -- read data from pipe
-function session:_read_pipe(opt)
+function server_session:_read_pipe(opt)
     local buff = bytes(256)
     local rpipe = opt.rpipe
     local verbose = option.get("verbose")
@@ -331,7 +326,7 @@ function session:_read_pipe(opt)
 end
 
 -- recv data from stream
-function session:_recv_data(buff)
+function server_session:_recv_data(buff)
     local stream = self:stream()
     local msg = stream:recv_msg()
     if msg and msg:is_data() then
@@ -340,7 +335,7 @@ function session:_recv_data(buff)
 end
 
 -- send data to stream
-function session:_send_data(data)
+function server_session:_send_data(data)
     local stream = self:stream()
     if stream:send_msg(message.new_data(self:id(), data:size())) then
         if stream:send(data) then
@@ -349,12 +344,30 @@ function session:_send_data(data)
     end
 end
 
-function session:__tostring()
+-- recv syncfiles
+function server_session:_recv_syncfiles(manifest, outputdir)
+    local stream = self:stream()
+    for _, fileitem in ipairs(manifest.inserted) do
+        local filepath = path.join(outputdir, fileitem)
+        if not stream:recv_file(filepath) then
+            return false
+        end
+    end
+    for _, fileitem in ipairs(manifest.modified) do
+        local filepath = path.join(outputdir, fileitem)
+        if not stream:recv_file(filepath) then
+            return false
+        end
+    end
+    return true
+end
+
+function server_session:__tostring()
     return string.format("<session %s>", self:id())
 end
 
-function main(session_id)
-    local instance = session()
-    instance:init(session_id)
+function main(server, session_id)
+    local instance = server_session()
+    instance:init(server, session_id)
     return instance
 end