+++ xmake/modules/private/service/server_config.lua
@@ -15,7 +15,7 @@
 -- Copyright (C) 2015-present, TBOOX Open Source Group.
 --
 -- @author      ruki
--- @file        config.lua
+-- @file        server_config.lua
 --
 
 -- imports
@@ -32,40 +32,38 @@ end
 function _generate_configfile()
     local filepath = configfile()
     assert(not _g.configs and not os.isfile(filepath))
-    local servicedir = path.join(global.directory(), "service")
+    local servicedir = path.join(global.directory(), "service", "server")
     local token = _generate_token()
     print("generating the config file to %s ..", filepath)
     cprint("an token(${bright yellow}%s${clear}) is generated, we can use this token to connect service.", token)
     local configs = {
         logfile = path.join(servicedir, "logs.txt"),
-        server = {
-            known_hosts = {
-            --    "127.0.0.1"
-            },
-            tokens = {
-                token
-            }
+        known_hosts = {
+        --    "127.0.0.1"
+        },
+        tokens = {
+            token
         },
         remote_build = {
-            server = {
-                listen = "0.0.0.0:9691",
-                workdir = path.join(servicedir, "remote_build"),
-            },
-            client = {
-                -- without authorization: "127.0.0.1:9691"
-                -- with user authorization: "user@127.0.0.1:9691"
-                connect = "127.0.0.1:9691",
-                -- with token authorization
-                token = token
+            listen = "0.0.0.0:9691",
+            workdir = path.join(servicedir, "remote_build"),
+        },
+        distcc_build = {
+            listen = "0.0.0.0:9692",
+            workdir = path.join(servicedir, "distcc_build"),
+            toolchains = {
+                ndk = {
+                }
             }
         }
+
     }
     save(configs)
 end
 
 -- get config file path
 function configfile()
-    return path.join(global.directory(), "service.conf")
+    return path.join(global.directory(), "service", "server.conf")
 end
 
 -- get all configs
@@ -73,7 +71,7 @@ function configs()
     return _g.configs
 end
 
--- get the given config, e.g. config.get("remote_build.server.listen")
+-- get the given config, e.g. server_config.get("remote_build.listen")
 function get(name)
     local value = configs()
     for _, key in ipairs(name:split('.', {plain = true})) do