+++ xmake/modules/private/service/stream.lua
@@ -160,6 +160,11 @@ function stream:send_string(str, opt)
     return self:send_data(bytes(str), opt)
 end
 
+-- send empty data
+function stream:send_emptydata(opt)
+    return self:send_header(0)
+end
+
 -- send file
 function stream:send_file(filepath, opt)
 
@@ -167,6 +172,7 @@ function stream:send_file(filepath, opt)
     opt = opt or {}
     local flags = 0
     local tmpfile
+    local originsize = os.filesize(filepath)
     if opt.compress then
         flags = bit.bor(flags, HEADER_FLAG_COMPRESS_LZ4)
         tmpfile = os.tmpfile()
@@ -199,7 +205,25 @@ function stream:send_file(filepath, opt)
     if tmpfile then
         os.tryrm(tmpfile)
     end
-    return ok
+    if ok then
+        return originsize, size
+    end
+end
+
+-- send files
+function stream:send_files(filepaths, opt)
+    local size
+    local compressed_size
+    for _, filepath in ipairs(filepaths) do
+        local real, compressed_real = self:send_file(filepath, opt)
+        if real then
+            size = (size or 0) + real
+            compressed_size = (compressed_size or 0) + compressed_real
+        else
+            return
+        end
+    end
+    return size, compressed_size
 end
 
 -- recv the given bytes
@@ -328,10 +352,8 @@ end
 function stream:recv_file(filepath)
     local size, flags = self:recv_header()
     if size then
-        local dstfile
         if bit.band(flags, HEADER_FLAG_COMPRESS_LZ4) == HEADER_FLAG_COMPRESS_LZ4 then
-            dstfile = filepath
-            filepath = os.tmpfile()
+            return self:_recv_compressed_file(lz4.decompress_stream(), filepath, size)
         end
         local buff = self._BUFF
         local recv = 0
@@ -345,12 +367,59 @@ function stream:recv_file(filepath)
         end
         file:close()
         if recv == size then
-            if dstfile then
-                lz4.decompress_file(filepath, dstfile)
-                os.tryrm(filepath)
-            end
-            return true
+            return recv
         end
+    end
+end
+
+-- recv files
+function stream:recv_files(filepaths)
+    local size, decompressed_size
+    for _, filepath in ipairs(filepaths) do
+        local real, decompressed_real = self:recv_file(filepath)
+        if real then
+            size = (size or 0) + real
+            decompressed_size = (decompressed_size or 0) + decompressed_real
+        else
+            return
+        end
+    end
+    return size, decompressed_size
+end
+
+-- recv compressed file
+function stream:_recv_compressed_file(lz4_stream, filepath, size)
+    local buff = self._BUFF
+    local recv = 0
+    local file = io.open(filepath, "wb")
+    local decompressed_size = 0
+    while recv < size do
+        local data = self:recv(buff, math.min(buff:size(), size - recv))
+        if data then
+            local write = 0
+            local writesize = data:size()
+            while write < writesize do
+                local blocksize = math.min(writesize - write, 8192)
+                local real = lz4_stream:write(data, {start = write + 1, last = write + blocksize})
+                if real > 0 then
+                    while true do
+                        local decompressed_real, decompressed_data = lz4_stream:read(buff, 8192)
+                        if decompressed_real > 0 and decompressed_data then
+                            file:write(decompressed_data)
+                            decompressed_size = decompressed_size + decompressed_real
+                        else
+                            break
+                        end
+                    end
+                end
+                write = write + blocksize
+            end
+            recv = recv + data:size()
+        end
+    end
+    file:close()
+    if recv == size then
+        return recv, decompressed_size
     end
 end