+++ CHANGELOG.md
@@ -8,10 +8,16 @@
 * Add path instance interfaces
 * [#2334](https://github.com/xmake-io/xmake/pull/2334): Add lz4 compress module
 * [#2349](https://github.com/xmake-io/xmake/pull/2349): Add keil/c51 project support
+* [#274](https://github.com/xmake-io/xmake/issues/274): Distributed compilation support
 
 ### Changes
 
 * [#2309](https://github.com/xmake-io/xmake/issues/2309): Support user authorization for remote compilation
+* Improve remote compilation to support lz4 compression
+
+### Bugs fixed
+
+* Fix lua stack when select package versions
 
 ## v2.6.5
 
@@ -1263,10 +1269,16 @@
 * 添加 path 实例接口
 * [#2334](https://github.com/xmake-io/xmake/pull/2334): 添加 lz4 压缩模块
 * [#2349](https://github.com/xmake-io/xmake/pull/2349): 添加 keil/c51 工程支持
+* [#274](https://github.com/xmake-io/xmake/issues/274): 跨平台分布式编译支持
 
 ### 改进
 
 * [#2309](https://github.com/xmake-io/xmake/issues/2309): 远程编译支持用户授权验证
+* 改进远程编译，增加对 lz4 压缩支持
+
+### Bugs 修复
+
+* 修复选择包版本时候 lua 栈不平衡导致的崩溃问题
 
 ## v2.6.5